requirements.txt
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.4
DateTime==5.5
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
pytz==2024.2
requests==2.32.3
setuptools==75.2.0
urllib3==2.2.3
zope.interface==7.1.1
